requirements.txt
alembic
ariadne
amqp
aniso8601
appdirs
asgiref
bcrypt
billiard
black
celery
cffi
click
click-didyoumean
click-plugins
click-repl
colorama
fastapi
flower
graphene==3.0.0b7  # Beta version that supports graphql-core 3
graphene-sqlalchemy==3.0.0b1
graphql-core==3.1.7
graphql-relay==3.0.0
greenlet
h11
humanize
kombu
Mako
MarkupSafe
mypy-extensions
pathspec
prometheus-client
promise
prompt-toolkit
psycopg2-binary
pycparser
pydantic
PyJWT
python-dateutil
python-dotenv
python-editor
pytz
redis
regex
Rx
singledispatch
six
SQLAlchemy
starlette
starlette-graphql
starlette_graphene3
toml
tornado
typing-extensions
uvicorn
vine
wcwidth